"added varibles for the map"
