+++ assets/style.css
@@ -3,11 +3,6 @@
     height: 500px;
     width: 500px;
 }
-/* html, body{
-    height: 100%;
-    margin: 0px;
-    padding: 0px;
-} */
 
 'colors': {
     '50': '#f6f6f7',
