+++ index.html
@@ -25,7 +25,7 @@
 
         <!-- Main / center content here -->
         <main class="col-span-8 bg-gray-500">
-            Main Content
+            <div id="map"></div>
         </main>
 
         <!-- Right sidebar - aside -->
